# core/docs/swagger.yaml
basePath: /api/v1
definitions:
...
  controller.WikiImageUploadResponse:
    properties:
      height:
        type: integer
      id:
        type: string
      url:
        type: string
      width:
        type: integer
    type: object
...
  responses.ErrorResponse:
    properties:
      error:
        type: string
    type: object
...
info:
  contact: {}
  description: Command-and-control core backend API.
  title: VibeC2 Core API
  version: "1.0"
paths:
...
  /wiki/images:
    post:
      consumes:
      - multipart/form-data
      description: Upload an image to be embedded in a wiki document. Images are validated,
        sanitized (SVG) or re-encoded (raster, strips EXIF), and optionally downscaled.
      parameters:
      - description: Owning document ID (UUID)
        in: formData
        name: documentId
        required: true
        type: string
      - description: Image bytes (png/jpeg/webp/gif/avif/svg, max 10MB)
        in: formData
        name: file
        required: true
        type: file
      produces:
      - application/json
      responses:
        "201":
          description: Created
          schema:
            $ref: '#/definitions/controller.WikiImageUploadResponse'
        "400":
          description: Bad Request
          schema:
            $ref: '#/definitions/responses.ErrorResponse'
        "403":
          description: Forbidden
          schema:
            $ref: '#/definitions/responses.ErrorResponse'
        "404":
          description: Not Found
          schema:
            $ref: '#/definitions/responses.ErrorResponse'
        "413":
          description: Request Entity Too Large
          schema:
            $ref: '#/definitions/responses.ErrorResponse'
        "415":
          description: Unsupported Media Type
          schema:
            $ref: '#/definitions/responses.ErrorResponse'
      security:
      - BearerAuth: []
      summary: Upload a wiki image
      tags:
      - Wiki
  /wiki/images/{id}:
    get:
      description: Streams the image bytes for an uploaded wiki image. Caller must
        be a member of the image's operation.
      parameters:
      - description: Image ID (UUID)
        in: path
        name: id
        required: true
        type: string
      produces:
      - image/*
      responses: {}
      security:
      - BearerAuth: []
      summary: Fetch a wiki image
      tags:
      - Wiki
...
securityDefinitions:
  BearerAuth:
    description: Bearer JWT token (e.g. "Bearer eyJhb...")
    in: header
    name: Authorization
    type: apiKey
swagger: "2.0"